fix: update types, constants
