+++ src/components/complete-profile-second/work-form/constants.ts
@@ -2,8 +2,6 @@ import { CompleteProfileSecondValues } from 'src/components/complete-profile-sec
 
 export const DATE_FORMAT = 'dd/MM/yyyy';
 
-export const MAX_CHARACTERS_LENGTH = 100;
-
 export const DEFAULT_EXPERIENCE_VALUES: CompleteProfileSecondValues = {
   workPlace: '',
   workType: '',
