+++ src/components/complete-profile-second/work-form/validation.ts
@@ -1,6 +1,6 @@
 import { ObjectSchema, boolean, date, object, string } from 'yup';
 import { useLocales } from 'src/locales';
-import { MAX_CHARACTERS_LENGTH } from './constants';
+import { MAX_CHARACTERS_LENGTH } from 'src/constants';
 import { CompleteProfileSecondValues } from '../types';
 
 export const useProfileExperienceSchema = (): ObjectSchema<CompleteProfileSecondValues> => {
