+++ src/constants/index.ts
@@ -4,6 +4,7 @@ export const EMAIL_ERROR = 'email';
 export const PHONE_ERROR = 'phone';
 export const USER_MIN_AGE = 18;
 export const USER_DATE_BIRTH_FORMAT = 'dd.MM.yyyy';
+export const MAX_CHARACTERS_LENGTH = 100;
 
 export const USER_ROLE = {
   caregiver: 'caregiver',
